Fix millis() rollover bug in obstruction sensor timing
The previous code `current_millis - last_asleep > 700` would fail when millis()
rolls over after ~49 days. When last_asleep is from before rollover and
current_millis is after, the subtraction produces a very large positive number
instead of the small duration, causing false obstruction detection.

The cast to (unsigned long) ensures proper rollover-safe arithmetic per
Arduino conventions.

+++ src/ratgdo.cpp
@@ -314,7 +314,7 @@ void obstruction_timer()
             else
             {
                 // if the line is high and was last asleep more than 700ms ago, then there is an obstruction present
-                if (current_millis - last_asleep > 700)
+                if ((unsigned long)(current_millis - last_asleep) > 700)
                 {
                     // Only update if we are changing state
                     if (!garage_door.obstructed)
